backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
alembic>=1.7.1,<2.0.0
python-dotenv>=0.19.0,<0.20.0
pandas>=1.3.0,<2.0.0
openpyxl>=3.0.7,<4.0.0
redis>=4.0.0,<5.0.0
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.18.0,<0.19.0
httpx>=0.24.0,<0.25.0
pytest-cov>=4.0.0,<5.0.0
prometheus-client>=0.16.0,<0.17.0
psutil>=5.9.0,<6.0.0